no se ya me olvide, cambios a empleados y a pruebas. Implementacion de EmpleadoDTO en el modelo de prueba

+++ empleados/src/main/java/utn/frc/backend/empleados/model/Empleado.java
@@ -27,5 +27,5 @@ public class Empleado {
     @Column(name = "APELLIDO")
     private String apellido;
     @Column(name = "TELEFONO_CONTACTO")
-    private int telefonoContacto;
+    private long telefonoContacto;
 }
